fix rejects
