+++ internal/provider/resource_dashboard_test.go
@@ -11,6 +11,27 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
+func TestAccDashboardResource_Simple(t *testing.T) {
+	timestamp := time.Now().Unix()
+	dashboardName := fmt.Sprintf("simple_dashboard_%d", timestamp)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: testAccDashboardResourceConfigSimple(dashboardName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("groundcover_dashboard.test", "name", dashboardName),
+					resource.TestCheckResourceAttrSet("groundcover_dashboard.test", "id"),
+					resource.TestCheckResourceAttrSet("groundcover_dashboard.test", "revision_number"),
+				),
+			},
+		},
+	})
+}
+
 func TestAccDashboardResource(t *testing.T) {
 	timestamp := time.Now().Unix()
 	dashboardName := fmt.Sprintf("test_dashboard_%d", timestamp)
@@ -197,6 +218,22 @@ resource "groundcover_dashboard" "test" {
         }
       }
     ]
+    variables     = {}
+    schemaVersion = 3
+  })
+}
+`, name)
+}
+
+func testAccDashboardResourceConfigSimple(name string) string {
+	return fmt.Sprintf(`
+resource "groundcover_dashboard" "test" {
+  name        = "%s"
+  description = "Simple test dashboard"
+  preset      = jsonencode({
+    duration      = "Last 1 hour"
+    widgets       = []
+    layout        = []
     variables     = {}
     schemaVersion = 3
   })
